requirements.txt
llama-index
llama-index-embeddings-huggingface
llama-index-llms-google-genai
llama-index-vector-stores-chroma
chromadb
python-dotenv
sentence-transformers
pypdf
python-docx
tqdm
docx2txt


streamlit
rank-bm25
torch
transformers

